Submission of Exercises 4.20.-4.22.

+++ controllers/blogs.js
@@ -2,34 +2,22 @@ const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
-const getTokenFrom = req => {
-    const authorization = req.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
 
 blogsRouter.get('/', async (req, res) => {
     const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
     res.json(blogs)
 })
   
-blogsRouter.post('/', async (req, res) => {
-    console.log(getTokenFrom(req), "gettokenform")
-    const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET)
-    // console.log(decodedToken, "decodedtoken")
-
+blogsRouter.post('/', userExtractor, async (req, res) => {
+    
     if (!req.body.title || !req.body.url) {
         return res.status(400).json({error: 'content missing'})
     }
 
-    if (!decodedToken.id) {
-        return res.status(401).json({ error: 'invalid token' })
-    }
-
-    const user = await User.findById(decodedToken.id)
+    const user = req.user
+    console.log(user)
 
     const blogV2 = {
         title: req.body.title,
@@ -47,7 +35,20 @@ blogsRouter.post('/', async (req, res) => {
     res.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', userExtractor, async (req, res) => {
+
+    const user = req.user
+    const blog = await Blog.findById(req.params.id)
+
+    console.log("user", user, "enduser")
+
+    const updatedBlogs = user.blogs.filter(blogid => blogid.toString() !== blog.id)
+
+    if (blog.user.toString() !== user.id) {
+        return res.status(401).json({ error: 'invalid user or token' })
+    } 
+
+    await User.findByIdAndUpdate(user.id, {blogs: updatedBlogs}, {new: true, runValidators: true})
     await Blog.findByIdAndDelete(req.params.id)
     res.status(204).end()
 })
@@ -60,7 +61,7 @@ blogsRouter.put('/:id', async (req, res) => {
         likes: req.body.likes,
     }
 
-    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blogInReq, {new: true})
+    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blogInReq, {new: true, runValidators: true})
     res.json(updatedBlog)
 })
